refactor: Concert id 필드 타입 변경

+++ src/main/java/org/example/interpark/domain/concert/dto/response/ConcertSearchResponseDto.java
@@ -1,5 +1,5 @@
 package org.example.interpark.domain.concert.dto.response;
 
-public record ConcertSearchResponseDto(Long id, String name, int amount) {
+public record ConcertSearchResponseDto(int id, String name, int amount) {
 
 }
